+++ server/src/AuthService/Seeders/RoleSeeder.cs
@@ -1,0 +1,31 @@
+using AuthService.Data;
+using Microsoft.AspNetCore.Identity;
+
+namespace AuthService.Seeders;
+
+public class RoleSeeder : BackgroundService
+{
+    private readonly IServiceProvider _serviceProvider;
+
+    public RoleSeeder(IServiceProvider serviceProvider) => _serviceProvider = serviceProvider;
+
+    protected override async Task ExecuteAsync(CancellationToken cancellationToken)
+    {
+        await using var scope = _serviceProvider.CreateAsyncScope();
+
+        var context = scope.ServiceProvider.GetRequiredService<AuthDbContext>();
+        await context.Database.EnsureCreatedAsync(cancellationToken);
+
+        var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole>>();
+
+        string[] roleNames = ["admin", "guest"];
+        foreach (var roleName in roleNames)
+        {
+            var roleExist = await roleManager.RoleExistsAsync(roleName);
+            if (!roleExist)
+            {
+                await roleManager.CreateAsync(new IdentityRole(roleName));
+            }
+        }
+    }
+}
